workshop/railway_deployment/hook_boost_railway/requirements.txt
discord.py==2.3.2
requests==2.31.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
python-dateutil==2.8.2
pytz==2023.3
schedule==1.2.0 